+++ paywall/paywall-builder/mutationObserver.js
