+++ paywall/src/__tests__/paywall-builder/index.test.js
@@ -1,4 +1,4 @@
-import * as mutations from '../../../paywall-builder/mutationobserver'
+import * as mutations from '../../../paywall-builder/mutationObserver'
 import * as buildManager from '../../../paywall-builder/build'
 import * as blockerManager from '../../../paywall-builder/blocker'
