+++ unlock-app/src/__tests__/paywall-builder/mutationObserver.test.js
@@ -1,7 +1,7 @@
 import {
   listenForNewLocks,
   changeListener,
-} from '../../paywall-builder/mutationobserver'
+} from '../../paywall-builder/mutationObserver'
 
 function tag(tagName, name, content) {
   return { nodeName: tagName.toUpperCase(), name, content }
