+++ unlock-app/src/paywall-builder/mutationObserver.js
